+++ src/components/common/TodoList/DraggableTodoList.tsx
@@ -6,7 +6,7 @@ import {
   Todo,
   // TodoUpdateRequest
 } from '@/shared/types/todo';
-import TodoListItem from '@/components/common/TodoListItem/unsafe_TodoListItem';
+import TodoListItem from '@/components/common/TodoListItem/index';
 import { Container } from './styles';
 
 interface TodoListProps {
